requirements.txt
Flask
Flask-RESTful
Flask-SQLAlchemy
mysqlclient
psycopg2-binary
PyJWT
gunicorn